test: address second round of Copilot review comments on PR #172
- concisenessscoring: remove scorerOnce.Do() from cleanup so loadScorer
  can safely re-initialize when origScorer is nil
- headingstyle: fold explore3_test.go into rule_morecoverage_test.go and
  delete the ad-hoc exploration file
- plan 85: remove unchecked mutation-testing criterion so ✅ status is
  consistent with all acceptance criteria being met
